item admin ,securing the views

+++ items/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 
 from .models import Item
@@ -46,10 +47,15 @@ def item_detail(request, item_id):
     return render(request, 'items/item_detail.html', context)
 
 
+@login_required
 def adding_item(request):
     """
     Adding items to the webstore
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES)
         if form.is_valid():
@@ -69,10 +75,15 @@ def adding_item(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_item(request, item_id):
     """
     Editing items in the webstore
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse('home'))
+
     item = get_object_or_404(Item, pk=item_id)
     if request.method == 'POST':
         form = ItemForm(request.POST, request.FILES, instance=item)
@@ -95,11 +106,16 @@ def edit_item(request, item_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_item(request, item_id):
 
     """
     Delete item from the webstore
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store owners can do that')
+        return redirect(reverse('home'))
+
     item = get_object_or_404(Item, pk=item_id)
     item.delete()
     messages.success(request, 'Item was successfully deleted')
